Multi resource handlers (#54)
* Implement support for multiResourceHandlers

A multi-resource handler is one that can process more than one
resource at a time. This could matter if an API has a batch add
function for example, in which case the Handler needs to receive
all the resources in one go. Or, if the handler accepts more than
one type of resource that have a hierarchical nature, e.g.
dashboard folders and dashboards.

To achieve this, a Handler takes responsibility for the whole
apply/diff process across those resources.

* Fix preview output

* JsonPath->JsonPaths

* Rename Path->JSONPath