Add config "get" and "path" subcommands (#329)
- get: Allows quickly getting the current config without going into the YAML settings file
- path: Allows actually finding the current config file. Had trouble finding the `/Users/julienduchesne/Library/Application Support/grizzly/settings.yaml` on my machine. This helps

While implementing the `config path` function, I noticed that the config logic was a bit weird. While grizzly was possibly reading (through viper) the config file from the workdir, it was writing back to the system config. Is this expected? By using `viper.ConfigFileUsed()`, we write back to the file we're reading the config from